Extracts MonitorMapFeature and MonitorMarkerProperties to monitor types
